Adds OpenAI batch upload logic

+++ SpotifyScraper/getTrackDetails.py
@@ -3,52 +3,80 @@ import json
 import pprint
 import time
 import base64
-from config import *
 import random
+import os
+from dotenv import load_dotenv
+import re
+from . import utils
 
-client_id = CLIENT_ID_1
-client_secret = CLIENT_SECRET_1
-access_token = ACTOKEN_1
-refresh_token = RFTOKEN_1
+load_dotenv()
+global access_token
+client_id = os.environ.get("CLIENT_ID_3")
+client_secret = os.environ.get("CLIENT_SECRET_3")
 
 
 def updateComposerTrackDetails():
-    with open(r'C:\Users\leonardo.calasans\Desktop\MusicPearls\data\composers.json', encoding='utf-8') as fp:
+    """
+    Fetches details of all tracks in data/tracks
+    """
+    global access_token
+    access_token = None
+    script_dir = os.path.dirname(os.path.abspath(__file__))
+
+    # Get all composers
+    with open(os.path.join(script_dir, '../data/composers.json'), encoding='utf-8') as fp:
         composers = json.load(fp)
     composers = composers['composers']
 
-    for i in range(11, 12):
-        print('------------------------------------------')
-        print(f'Fetching track details of {composers[i]["name"]}')
-        print('------------------------------------------')
+    # Get all existing track details files
+    files = os.listdir(os.path.join(script_dir, '../data/tracksDetails'))
+    pattern = re.compile(r'tracksDetails_(.+)\.json')
+    existingIds = [pattern.match(file).group(1) for file in files if pattern.match(file)]
 
-        composerTracksDetails = getComposerTrackDetails(composers[i]["id"])
-        composerTracksDetailsJson = {'tracks': composerTracksDetails}
-
-        with open(f'tracksDetails2/tracksDetails_{composers[i]["id"]}.json', 'w', encoding='utf-8') as f:
-            json.dump(composerTracksDetailsJson, f, ensure_ascii=False)
-        time.sleep(5 * (random.random() + 1))
-
+    i = 0
+    while i < len(composers):
+        composerName = composers[i]['name']
+        composerId = composers[i]['id']
+        if composerId in existingIds:
+            print(f'Skipping tracksDetails fetch of {composerName}')
+            i = i + 1
+        elif not os.path.exists(os.path.join(script_dir, f"../data/tracks/tracks_{composerId}.json")):
+            print(f'Tracks of {composerName} not found. Skipping')
+            i = i + 1
+        else:
+            print('------------------------------------------')
+            print(f'Fetching track details of {composers[i]["name"]}')
+            print('------------------------------------------')
+            composerTracksDetails = getComposerTrackDetails(composerId)
+            uniqueTracksDetails = list({ each['id'] : each for each in composerTracksDetails }.values())
+            composerTracksDetailsJson = {'tracks': uniqueTracksDetails}
+            with open(os.path.join(script_dir, f'../data/tracksDetails/tracksDetails_{composerId}.json'), 'w', encoding='utf-8') as f:
+                json.dump(composerTracksDetailsJson, f, ensure_ascii=False)
+            time.sleep(5 * (random.random() + 1))
+            i = i + 1
 
 def getComposerTrackDetails(artistId):
-    # Open track id file and get all ids of tracks
-    with open(fr'C:\Users\leonardo.calasans\Desktop\MusicPearls\data\tracks2\tracks_{artistId}.json',
-              encoding='utf-8') as fp:
+    """
+    Fetches details of all tracks from given composer
+    
+    Returns:
+        list[object]: list with objects with name and id of track, composer, other artists, and popularity
+    """
+
+    global access_token
+    script_dir = os.path.dirname(os.path.abspath(__file__))
+    with open(os.path.join(script_dir, f"../data/tracks/tracks_{artistId}.json"), encoding='utf-8') as fp:
         tracks = json.load(fp)
     tracks = tracks['tracks']
-
-    # Initialize variables
     composerTracksDetails = []
-    chunkSize = 50
-    global access_token
-    url = 'https://api.spotify.com/v1/tracks?'
 
-    # Loop through all tracks getting their details
+    chunkSize = 50
+    url = 'https://api.spotify.com/v1/tracks?'
     j = 0
     while j < len(tracks):
         time.sleep(1 * (random.random() + 1))
 
-        print(f"Getting tracks {j} to {j + chunkSize}")
+        print(f"Getting tracks {j} to {j + chunkSize} from {artistId}")
         tracksList = tracks[j:j + chunkSize]
         tracksStr = ','.join(tracksList)
         fetch_url = f"{url}ids={tracksStr}"
@@ -57,7 +85,7 @@ def getComposerTrackDetails(artistId):
         # If not authorized, get new token
         if response.status_code == 401:
             print(f'Reponse 401. Getting new token')
-            access_token = getNewToken()
+            access_token = utils.getNewToken(client_id, client_secret)
 
         # If over-request, wait and retry afterwards
         elif response.status_code == 429:
@@ -84,21 +112,3 @@ def getComposerTrackDetails(artistId):
             print('Bad response, trying again')
 
     return composerTracksDetails
-
-
-def getNewToken():
-    url = 'https://accounts.spotify.com/api/token'
-    auth_client = client_id + ":" + client_secret
-    auth_encode = 'Basic ' + base64.b64encode(auth_client.encode()).decode()
-    response = requests.post(url,
-                             headers={'Content-Type': 'application/x-www-form-urlencoded',
-                                      'Authorization': auth_encode},
-                             data={'grant_type': 'refresh_token', 'refresh_token': refresh_token})
-    if response.status_code == 200:
-        responseJson = response.json()
-        return responseJson['access_token']
-    else:
-        raise RuntimeError('Unable to refresh Access Token')
-
-
-updateComposerTrackDetails()
